requirements.txt
flask
Flask-SQLAlchemy
flask-login
boto3
requests
requests-aws4auth
python-dotenv